+++ client/src/components/categorySelection/categorySelection.css
@@ -1,0 +1,6 @@
+fieldset {
+
+    padding-bottom: 20px;
+    margin: 20px;
+    text-align: left;
+}
