+++ client/src/pages/UploadPage/Upload.js
@@ -1,5 +1,9 @@
 import './Upload.css';
 import React, { useState } from 'react';
+import CategorySelection from '../../components/categorySelection/categorySelection';
+
+import IngredientsBox from '../../components/ingredientsBox/ingredientsBox';
+
 
 const Upload = () => {
     
@@ -7,14 +11,17 @@ const Upload = () => {
     const [selectedCategory, setCategory] = useState('');
     const [totalTime, setTotalTime] = useState('');
 
-    const [selectedIngredient, setSelectedIngredient] = useState('');
-    const [customIngredient, setCustomIngredient] = useState('');
-    const [selectedQuantity, setSelectedQuantity] = useState('');
-    const [selectedUnit, setSelectedUnit] = useState('');
-    const [customUnit, setCustomUnit] = useState('');
+    // const [selectedIngredient, setSelectedIngredient] = useState('');
+    // const [customIngredient, setCustomIngredient] = useState('');
+    // const [selectedQuantity, setSelectedQuantity] = useState('');
+    // const [selectedUnit, setSelectedUnit] = useState('');
+    // const [customUnit, setCustomUnit] = useState('');
+
     const [ingredientsList, setIngredientsList] = useState([]);
 
     const [instructions, setInstructions] = useState('');
+
+    const [image, setImage] = useState(null);
 
     const [error, setError] = useState('');
 
@@ -23,95 +30,100 @@ const imageUrl = '';
 const createdBy = 'Mika check';
 
 
-
     const handleCategoryChange = (event) => {
         setCategory(event.target.value); 
       };
 
-
-    const predefinedIngredients = ['Flour', 'Sugar', 'Butter', 'Eggs'];
-    const unitsList = ['grams', 'cups', 'tablespoons', 'teaspoons'];
-
-
-    const handleIngredientChange = async (event) => {
-        const value = event.target.value;
-        setSelectedIngredient(value);
-        if (value !== 'Other') {
-            setCustomIngredient('');
-        }
-    };
-
-    const handleCustomIngredientChange = async (event) => {
-        setCustomIngredient(event.target.value);
-    };
-
-    const handleQuantityChange = async (event) => {
-        const value = event.target.value;
-        if (value < 0) {
-            setError('Quantity cannot be negative');
-        } else if (value === '') {
-            setError('Quantity is required');
-        } else {
-            setError('');
-        }
-        setSelectedQuantity(value);
+      const handleImageChange = (e) => {
+        setImage(e.target.files[0]);
     };
 
 
-    const handleUnitChange = async (event) => {
-        const value = event.target.value;
-        setSelectedUnit(value);
-        if (value !== 'Other') {
-            setCustomUnit('');
-        }
-    };
-
-    const handleCustomUnitChange = async (event) => {
-        setCustomUnit(event.target.value);
-    };
+    // const predefinedIngredients = ['Flour', 'Sugar', 'Butter', 'Eggs'];
+    // const unitsList = ['grams', 'cups', 'tablespoons', 'teaspoons'];
 
 
-    const addIngredient = async () => {
-        if (!selectedIngredient && !customIngredient) {
-            setError('Ingredient is required');
-            return;
-        }
-        if (!selectedQuantity) {
-            setError('Quantity is required');
-            return;
-        }
-        if (!selectedUnit && !customUnit) {
-            setError('Unit is required');
-            return;
-        }
+    // const handleIngredientChange = async (event) => {
+    //     const value = event.target.value;
+    //     setSelectedIngredient(value);
+    //     if (value !== 'Other') {
+    //         setCustomIngredient('');
+    //     }
+    // };
 
-        const ingredient = selectedIngredient === 'Other' ? customIngredient : selectedIngredient;
-        const unit = selectedUnit === 'Other' ? customUnit : selectedUnit;
-        const newIngredient = { ingredient, quantity: selectedQuantity, unit};
-        setIngredientsList([...ingredientsList, newIngredient]);
+    // const handleCustomIngredientChange = async (event) => {
+    //     setCustomIngredient(event.target.value);
+    // };
 
-        // Reset fields
-        setSelectedIngredient('');
-        setCustomIngredient('');
-        setSelectedQuantity('');
-        setSelectedUnit('');
-        setCustomUnit('');
-        setError('');
-
-    };
+    // const handleQuantityChange = async (event) => {
+    //     const value = event.target.value;
+    //     if (value < 0) {
+    //         setError('Quantity cannot be negative');
+    //     } else if (value === '') {
+    //         setError('Quantity is required');
+    //     } else {
+    //         setError('');
+    //     }
+    //     setSelectedQuantity(value);
+    // };
 
 
-    const removeIngredient = (index) => {
-        const newList = ingredientsList.filter((_, i) => i !== index);
-        setIngredientsList(newList);
-    };
+    // const handleUnitChange = async (event) => {
+    //     const value = event.target.value;
+    //     setSelectedUnit(value);
+    //     if (value !== 'Other') {
+    //         setCustomUnit('');
+    //     }
+    // };
+
+    // const handleCustomUnitChange = async (event) => {
+    //     setCustomUnit(event.target.value);
+    // };
+
+
+    // const addIngredient = async () => {
+    //     if (!selectedIngredient && !customIngredient) {
+    //         setError('Ingredient is required');
+    //         return;
+    //     }
+    //     if (!selectedQuantity) {
+    //         setError('Quantity is required');
+    //         return;
+    //     }
+    //     if (!selectedUnit && !customUnit) {
+    //         setError('Unit is required');
+    //         return;
+    //     }
+
+    //     const ingredient = selectedIngredient === 'Other' ? customIngredient : selectedIngredient;
+    //     const unit = selectedUnit === 'Other' ? customUnit : selectedUnit;
+    //     const newIngredient = { ingredient, quantity: selectedQuantity, unit};
+    //     setIngredientsList([...ingredientsList, newIngredient]);
+
+    //     // Reset fields
+    //     setSelectedIngredient('');
+    //     setCustomIngredient('');
+    //     setSelectedQuantity('');
+    //     setSelectedUnit('');
+    //     setCustomUnit('');
+    //     setError('');
+
+    // };
+
+
+    // const removeIngredient = (index) => {
+    //     const newList = ingredientsList.filter((_, i) => i !== index);
+    //     setIngredientsList(newList);
+    // };
 
 
     //edit this function
     const handleSubmit = async (e) => {
         e.preventDefault();
 
-        const recipe = {title, selectedCategory, imageUrl, ingredientsList, totalTime, instructions, createdBy}
+        const recipe = 
+        
+        {title, selectedCategory, imageUrl, ingredientsList, totalTime, instructions, createdBy}
 
        
         // const form = e.target;
@@ -119,7 +131,6 @@ const createdBy = 'Mika check';
 
         const response = await fetch('/api/upload', {
             method: 'POST',
-            // method: form.method,
             body: JSON.stringify(recipe),
             headers:{
                 'Content-Type': 'application/json'
@@ -134,11 +145,11 @@ const createdBy = 'Mika check';
             setCategory('')
             setTitle('')
             setTotalTime('')
-            setSelectedIngredient('')
-            setCustomIngredient('')
-            setSelectedQuantity('')
-            setSelectedUnit('')
-            setCustomUnit('')
+            // setSelectedIngredient('')
+            // setCustomIngredient('')
+            // setSelectedQuantity('')
+            // setSelectedUnit('')
+            // setCustomUnit('')
             setIngredientsList([])
             setInstructions('')
 
@@ -157,13 +168,10 @@ const createdBy = 'Mika check';
 
 
 
-
-
     return (  
 
         <div>
 
-        
             
             <form className="form" onSubmit={handleSubmit}>
 
@@ -177,80 +185,12 @@ const createdBy = 'Mika check';
                 onChange = {(e) => setTitle(e.target.value)}
                 value = {title}
                 />
+            
 
-            <fieldset>
-                <legend>Select Category: </legend>
-            {/* <p>Select category:</p> */}
-                <label>
-                    <input
-                    type="radio"
-                    name="Category"
-                    value="Appetizer"
-                    checked={selectedCategory === "Appetizer"} // Set checked status
-                    onChange={handleCategoryChange}
-                    />
-                    Appetizer
-                </label>
-                <br/>
-                <label>
-                    <input
-                    type="radio"
-                    name="Category"
-                    value="Soup"
-                    checked={selectedCategory === "Soup"}
-                    onChange={handleCategoryChange}
-                    />
-                    Soup
-                </label>
-                <br/>
-                <label>
-                    <input
-                    type="radio"
-                    name="Category"
-                    value="Salad"
-                    checked={selectedCategory === "Salad"}
-                    onChange={handleCategoryChange}
-                    />
-                    Salad
-                </label>
-                <br/>
-                <label>
-                    <input
-                    type="radio"
-                    name="Category"
-                    value="Main Course"
-                    checked={selectedCategory === "Main Course"}
-                    onChange={handleCategoryChange}
-                    />
-                    Main Course
-                </label>
-                <br/>
-                <label>
-                    <input
-                    type="radio"
-                    name="Category"
-                    value="Side Dish"
-                    checked={selectedCategory === "Side Dish"}
-                    onChange={handleCategoryChange}
-                    />
-                    Side Dish
-                </label>
-                <br/>
-                <label>
-                    <input
-                    type="radio"
-                    name="Category"
-                    value="Dessert"
-                    checked={selectedCategory === "Dessert"}
-                    onChange={handleCategoryChange}
-                    />
-                    Dessert
-                </label>
-                
-                {/* Display the selected category */}
-                <p>Selected Category: {selectedCategory}</p>
-                </fieldset>
-
+                < CategorySelection 
+                selectedCategory={selectedCategory}
+                handleCategoryChange={handleCategoryChange}
+                />
 
 
                 <div>
@@ -264,11 +204,24 @@ const createdBy = 'Mika check';
                 />
                 </div>
 
-                <p className='ingredientsContainer'>
 
-                {/* <div className="ingredientsContainer"> */}
+                <IngredientsBox
+                ingredientsList={ingredientsList}
+                setIngredientsList={setIngredientsList}
+                error={error}
+                setError={setError}
+                />
+
+
+
+
+
+
+                {/* <p className='ingredientsContainer'>
+
                 
                 <label>Ingredients List: </label>
+                <div className='choiceContainer'> 
                 <select placeholder="Ingredient " label="Ingredient" value={selectedIngredient} onChange={handleIngredientChange} className='selectBox'>
                     <option style={{color: 'red'}} value="" disabled> Select Ingredient </option>
                     {predefinedIngredients.map((ingredient) => (
@@ -295,7 +248,7 @@ const createdBy = 'Mika check';
                     onChange={handleQuantityChange}
                 />
  
-                {/* <label>Select Unit: </label> */}
+                
                 <select value={selectedUnit} onChange={handleUnitChange} className='selectBox'>
                     <option value="" disabled> Select Unit</option>
                     {unitsList.map((unit) => (
@@ -315,8 +268,9 @@ const createdBy = 'Mika check';
                 <button onClick={addIngredient}>Add Ingredient</button>
                 {error && <p>{error}</p>}
 
-                {/* </div> */}
-
+                
+                </div>
+                
                 <div className='ingredientsList'>
                     {ingredientsList.map((item, index) => (
                         <div className='ingredientItem'>    
@@ -327,20 +281,39 @@ const createdBy = 'Mika check';
                     ))}
                 </div>
                 </p>
-       
-                <textarea id="Instructions">
-                {/* <label>Instructions: </label> */}
-                    <input 
-                    type="text"
-                    className='instructionBox'
-                    name="Instructions"
-                    onChange = {(e) => setInstructions(e.target.value)}
-                    value = {instructions}
-                    />
-                </textarea>
+ */}
 
 
-                {/*needs to add an image alss*/}
+
+
+
+
+                <div className='instructionsBox'>
+                <label>Instructions: </label>
+                <input 
+                className='textInput'
+                type="text"
+                name = "Instructios"
+                onChange = {(e) => setInstructions(e.target.value)}
+                value = {instructions}
+                />
+                </div>
+
+
+
+                <div>
+                <label>Upload Image: </label>
+                <input 
+                    type="file"
+                    accept="image/*" // Restrict to image files only
+                    onChange={handleImageChange}
+                />
+                </div>
+
+
+
+
+
                 <hr />
                 <button type="reset">Reset form</button>
                 <button type="submit">Submit form</button>
