+++ server/package.json
@@ -19,6 +19,7 @@
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.8.1",
     "mongoose": "^8.5.1",
+    "multer": "^1.4.5-lts.1",
     "validator": "^13.12.0"
   },
   "devDependencies": {
